fix: 점수 결과 class명으로 나오는 bug fix
