+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -4,10 +4,10 @@ import blackjack.domain.GameBoard;
 import blackjack.domain.card.Deck;
 import blackjack.domain.card.RandomCardsShuffler;
 import blackjack.domain.participants.BettingMoney;
-import blackjack.domain.participants.BettingMoney.Profit;
 import blackjack.domain.participants.Dealer;
 import blackjack.domain.participants.Player;
 import blackjack.domain.participants.Players;
+import blackjack.domain.participants.Profit;
 import blackjack.domain.state.Created;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
