+++ src/main/java/blackjack/domain/participants/Profit.java
@@ -1,0 +1,13 @@
+package blackjack.domain.participants;
+
+import java.util.Map;
+
+public record Profit(
+        int dealerProfit,
+        Map<Player, Integer> playerProfit
+) {
+    public Profit(int dealerProfit, Map<Player, Integer> playerProfit) {
+        this.dealerProfit = dealerProfit;
+        this.playerProfit = playerProfit;
+    }
+}
